+++ app/index.js
@@ -56,8 +56,6 @@ require("./lib/jquery.format.js")
 let deparam = require("jquery-deparam")
 
 window.c = console
-// __IS_LAB__ is defined in webpack and set to true if NODE_ENV is "staging"
-window.isLab = __IS_LAB__
 window.currentMode = deparam(window.location.search.slice(1)).mode || "default"
 
 try {